Revert "Move the Local Artifact Services to the DirectRunner"
This reverts commit b5deb3175fcf85f528078cf7e151be7eec120d11.
