+++ runners/local-artifact-service-java/src/main/java/org/apache/beam/artifact/local/LocalFileSystemArtifactStagerService.java
@@ -1,25 +1,22 @@
 /*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
  *
- *  * Licensed to the Apache Software Foundation (ASF) under one
- *  * or more contributor license agreements.  See the NOTICE file
- *  * distributed with this work for additional information
- *  * regarding copyright ownership.  The ASF licenses this file
- *  * to you under the Apache License, Version 2.0 (the
- *  * "License"); you may not use this file except in compliance
- *  * with the License.  You may obtain a copy of the License at
- *  *
- *  *     http://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  * Unless required by applicable law or agreed to in writing, software
- *  * distributed under the License is distributed on an "AS IS" BASIS,
- *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  * See the License for the specific language governing permissions and
- *  * limitations under the License.
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
- *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
-package org.apache.beam.runners.direct.portable.artifact;
+package org.apache.beam.artifact.local;
 
 import static com.google.common.base.Preconditions.checkState;
 
@@ -48,7 +45,7 @@ public class LocalFileSystemArtifactStagerService
   private static final Logger LOG =
       LoggerFactory.getLogger(LocalFileSystemArtifactStagerService.class);
 
-  public static LocalFileSystemArtifactStagerService forRootDirectory(File base) {
+  public static LocalFileSystemArtifactStagerService withRootDirectory(File base) {
     return new LocalFileSystemArtifactStagerService(base);
   }
 
@@ -210,13 +207,9 @@ public class LocalFileSystemArtifactStagerService
     public void onNext(ArtifactApi.PutArtifactRequest value) {
       try {
         if (value.getData() == null) {
-          StatusRuntimeException e =
-              Status.INVALID_ARGUMENT
-                  .withDescription(
-                      String.format(
-                          "Expected all chunks in the current stream state to contain data, got %s",
-                          value.getContentCase()))
-                  .asRuntimeException();
+          StatusRuntimeException e = Status.INVALID_ARGUMENT.withDescription(String.format(
+              "Expected all chunks in the current stream state to contain data, got %s",
+              value.getContentCase())).asRuntimeException();
           throw e;
         }
         value.getData().getData().writeTo(target);
